added collisions

+++ src/physics.rs
@@ -1,6 +1,6 @@
 use bevy::{
     color::palettes::css::*,
-    math::Isometry2d,
+    math::bounding::{Aabb2d, BoundingVolume, IntersectsVolume},
     prelude::*,
 };
 
@@ -17,6 +17,7 @@ impl Plugin for PhysicsPlugin {
             (
                 apply_gravity,
                 apply_tongue_forces.run_if(frog_tongue_active),
+                check_collisions,
                 apply_force,
                 apply_velocity,
 
@@ -29,7 +30,9 @@ impl Plugin for PhysicsPlugin {
 #[derive(Debug, Component)]
 pub struct Rigidbody {
     pub mass: f32,
+    pub size: Vec2,
 
+    pub moveable: bool,
     pub velocity: Vec2,
     pub force: Vec2,
 }
@@ -38,7 +41,9 @@ impl Default for Rigidbody {
     fn default() -> Self {
         Self {
             mass: 1.0,
+            size: Vec2::new(20.0, 20.0),
 
+            moveable: true,
             velocity: Vec2::ZERO,
             force: Vec2::ZERO,
         }
@@ -52,6 +57,10 @@ fn apply_force(
     let delta = time.delta_secs();
 
     for mut rb in &mut query {
+        if !rb.moveable {
+            continue;
+        }
+
         if rb.mass <= 0.0 {
             continue;
         }
@@ -71,15 +80,17 @@ fn apply_velocity(
     let delta = time.delta_secs();
 
     for (mut transform, rb) in &mut query {
+        if !rb.moveable {
+            continue;
+        }
+
         transform.translation += (rb.velocity * delta).extend(0.0);
     }
 }
 
-fn apply_gravity(mut query: Query<(&mut Rigidbody, &Transform)>) {
-    for (mut rb, t) in &mut query {
-        // Hack solution
-        if t.translation.y < 0.0 {
-            rb.velocity.y = 0.0;
+fn apply_gravity(mut query: Query<&mut Rigidbody>) {
+    for mut rb in &mut query {
+        if !rb.moveable {
             continue;
         }
 
@@ -91,6 +102,10 @@ fn apply_tongue_forces(
     mut query: Query<(&mut Rigidbody, &mut Transform, &Frog)>,
 ) {
     for (mut rb, mut transform, frog) in &mut query {
+        if !rb.moveable {
+            continue;
+        }
+
         let current_pos = transform.translation.truncate();
         let offset = current_pos - frog.target_pos;
         let current_dist = offset.length();
@@ -117,18 +132,82 @@ fn apply_tongue_forces(
     }
 }
 
+fn check_collisions(mut query: Query<(Entity, &mut Transform, &mut Rigidbody)>) {
+    let mut combinations = query.iter_combinations_mut();
+
+    while let Some([(ent_a, mut t_a, mut rb_a), (ent_b, mut t_b, mut rb_b)]) = combinations.fetch_next() {
+        if !rb_a.moveable && !rb_b.moveable {
+            continue;
+        }
+
+        let aabb_a = Aabb2d::new(t_a.translation.truncate(), rb_a.size / 2.0);
+        let aabb_b = Aabb2d::new(t_b.translation.truncate(), rb_b.size / 2.0);
+
+        if aabb_a.intersects(&aabb_b) {
+            let a_hs = aabb_a.half_size();
+            let a_ctr = aabb_a.center();
+            let b_hs = aabb_b.half_size();
+            let b_ctr = aabb_b.center();
+
+            let overlap_x = (a_hs.x + b_hs.x) - (a_ctr.x - b_ctr.x).abs();
+            let overlap_y = (a_hs.y + b_hs.y) - (a_ctr.y - b_ctr.y).abs();
+
+            if overlap_x < overlap_y {
+                let sign = if a_ctr.x < b_ctr.x { -1.0 } else { 1.0 };
+                let push = Vec3::new(overlap_x * sign, 0.0, 0.0);
+
+                if rb_a.moveable && !rb_b.moveable {
+                    t_a.translation += push;
+                    rb_a.velocity.x = 0.0
+                } else if !rb_a.moveable && rb_b.moveable {
+                    t_b.translation -= push;
+                    rb_b.velocity.x = 0.0;
+                } else {
+                    t_a.translation += push * 0.5;
+                    t_b.translation -= push * 0.5;
+                    rb_a.velocity.x = 0.0;
+                    rb_b.velocity.x = 0.0;
+                }
+            } else {
+                let sign = if a_ctr.y < b_ctr.y { -1.0 } else { 1.0 };
+                let push = Vec3::new(0.0, overlap_y * sign, 0.0);
+                
+                if rb_a.moveable && !rb_b.moveable {
+                    t_a.translation += push;
+                    rb_a.velocity.y = 0.0
+                } else if !rb_a.moveable && rb_b.moveable {
+                    t_b.translation -= push;
+                    rb_b.velocity.y = 0.0;
+                } else {
+                    t_a.translation += push * 0.5;
+                    t_b.translation -= push * 0.5;
+                    rb_a.velocity.y = 0.0;
+                    rb_b.velocity.y = 0.0;
+                }
+            }
+
+        }
+    }
+}
+
 fn debug_gizmos(
     mut gizmos: Gizmos,
     query: Query<(&Transform, &Rigidbody)>
 ) {
-    gizmos.rect_2d(Isometry2d::IDENTITY, Vec2::splat(50.0), PINK);
 
-    // Draw velocity
     for (transform, rb) in query {
+        // Draw collider
+        if rb.moveable {
+            gizmos.rect_2d(transform.translation.truncate(), rb.size, PINK);
+        } else {
+            gizmos.rect_2d(transform.translation.truncate(), rb.size, RED);
+        }
+
+        // Draw velocity
         gizmos.arrow_2d(
             transform.translation.truncate(),
             transform.translation.truncate() + rb.velocity * 5.0,
             BLUE,
-        );    
+        );
     }
 }
